+++ changelog.md
@@ -3,6 +3,22 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/ru/1.0.0),
 and this project follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.1.3] - 2026-08-07
+
+### Added
+
+- created application config file - `resources/config.edn`
+- created application config getters - `core.config.*`
+
+### Changed
+
+- `core`, `domain.generator` now using properties from `config.edn`
+- code generation alphabet moved from `domain.generator` to `core.constants`
+
+### Fixed
+
+- fixed `domain.link` formatting
+
 ## [v0.1.2] - 2026-08-06
 
 ### Added